Move reusable code to a new folder in components
