+++ frontend/package.json
@@ -24,7 +24,7 @@
     "@mui/material": "^7.0.2",
     "@mui/system": "^7.0.2",
     "@mui/x-charts": "^8.1.0",
-    "@mui/x-data-grid": "^7.29.1",
+    "@mui/x-data-grid": "latest",
     "@nivo/bar": "^0.80.0",
     "@nivo/core": "^0.80.0",
     "@nivo/pie": "^0.80.0",
