+++ frontend/src/pages/VulnerabilityScanDash/TopVulnerabilities.tsx
@@ -22,14 +22,22 @@ type ColumnConfig<T> = {
 
 export default function TopVulnerabilities({
   allData,
-  kevData
+  kevData,
+  orgName,
+  orgId
 }: {
   allData: CVEItem[];
   kevData: CVEItem[];
+  orgName: string;
+  orgId: string;
 }) {
   const history = useHistory();
   const filteredVulnTableLinkHandler = (title: string) =>
-    history.push('/inventory/vulnerabilities', { title: title });
+    history.push('/inventory/vulnerabilities', {
+      title: title,
+      orgName,
+      orgId
+    });
 
   const [activeLabel, setActiveLabel] = useState('KEV');
   const handleTopVulnChipClick = (label: string) => {
