+++ frontend/src/pages/VulnerabilityScanDash/VulnerabilityScan.tsx
@@ -174,6 +174,14 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
       (currentOrganization as OrganizationShallow)?.id
     );
   }, [filters, currentOrganization]);
+  const organizationFilter = filters.find(
+    (filter) => filter.field === ORGANIZATION_FILTER_KEY
+  );
+
+  console.log('organizationFilter', organizationFilter?.values[0]?.name);
+  const orgName = organizationFilter?.values[0]?.name
+    ? organizationFilter?.values[0]?.name
+    : currentOrganization?.name;
 
   useEffect(() => {
     const selectedOrgId = orgId ? orgId : currentOrganization?.id;
@@ -433,6 +441,8 @@ export const VulnerabilityScan: React.FC<ContextType> = (props) => {
             <TopVulnerabilities
               allData={vulnScanData.topVulnerabilities}
               kevData={vulnScanData.topKevVulnerabilities}
+              orgName={orgName}
+              orgId={orgId}
             />
           </Grid>
         </Grid>
